새로운 투표 화면 완성
